refactor: split free start intake shell
